Add special case for CLES1 and CLES2; add two more channels to AKS
For AVO Winston, I've been using the get_availability() function to
figure out what special cases need to be hard-coded in (e.g. which
stations are BDF versus HDF). The AVO Winston special cases in this code
may need to be updated periodically, and get_availability() is a good
way to make sure that all such special cases are covered.