Implement automatic interpolation to lowest Trace fs present in Stream
Now, if decimation_rate is not specified, we will still equalize the
Trace sample rates to simplify downstream processing. (Previously, this
was only guaranteed if the user chose to decimate.) Note that now I'm
using the term "decimation" even though the ObsPy command employed is
actually interpolation. This change of wording better aligns with the
literature (e.g. Walker et al. (2010)).