Implement a custom RTMWarning class; update error messages
The justification for this is that RTM issues a lot of warnings so it's
a good idea for them to be more compact than the standard Python
warning format. Ultimately, the warning definition should be moved out
of warning_config.py and into __init__.py if/when we refactor this into
an actual Python package.