Implement DEM creation function
This should be generally useful for both topographically-influenced
propagation modeling as well as figure creation. Note that this function
requires GMT 6. Right now GMT 6 is in advanced dev status, but it should
be officially released later this summer. In any case, the current env
spec given in README.md should gather the correct GMT version.